Bug Fixes and Clean up code
Cold Iron now works. Clean up the code. Changes to Draconic Resistance to our version being Draconic Toughness with the AC change that comes with that
