+++ Javascript/initScript/weapon.mjs
@@ -3,17 +3,17 @@
 */
 
 export function weaponTotal(){
+    console.log("Elkan 5e  |  Initializing Weapons")
     weaponTypes()
     weaponProperties()
 }
 
 export function weaponProperties(){
-    CONFIG.DND5E.weaponProperties.cold = "Cold Iron"
-    CONFIG.DND5E.physicalWeaponProperties.cold = "Cold Iron"
+    CONFIG.DND5E.weaponProperties.coldIron = "Cold Iron"
+    CONFIG.DND5E.physicalWeaponProperties.coldIron = "Cold Iron"
     CONFIG.DND5E.weaponProperties.mou = "Mounted"
     CONFIG.DND5E.weaponProperties.unw = "Unwieldy"
     delete CONFIG.DND5E.weaponProperties.spc
-
 }
 
 export function weaponTypes(){
