--- Javascript/startingScripts/armor.js
@@ -1,23 +1,0 @@
-/*
-Changing Armor
-*/
-
-export function armor(){
-    CONFIG.DND5E.armorIds.breastplate = "elkan5e.elkan5e-mundane-items.47uWPwiJYyUXy3Al"
-    CONFIG.DND5E.armorIds.chainmail = "elkan5e.elkan5e-mundane-items.sgcCg8pGXKBu3f4S"
-    CONFIG.DND5E.armorIds.chainshirt = "elkan5e.elkan5e-mundane-items.rBW1VVkWQN2UYYrC"
-    CONFIG.DND5E.armorIds.halfplate = "elkan5e.elkan5e-mundane-items.PddUBElickanChFt"
-    CONFIG.DND5E.armorIds.hide = "elkan5e.elkan5e-mundane-items.O8vBudyTBZs03cTF"
-    CONFIG.DND5E.armorIds.leather = "elkan5e.elkan5e-mundane-items.0Kmj7Hk25ZyGURMI"
-    CONFIG.DND5E.armorIds.padded = "elkan5e.elkan5e-mundane-items.jSi9ralAm7FzUQbH"
-    CONFIG.DND5E.armorIds.plate = "elkan5e.elkan5e-mundane-items.k8m8rccrVYiIDiUC"
-    delete CONFIG.DND5E.armorIds.ringmail
-    CONFIG.DND5E.armorIds.scalemail = "elkan5e.elkan5e-mundane-items.QhHQy7whejLOVFf8"
-    CONFIG.DND5E.armorIds.splint = "elkan5e.elkan5e-mundane-items.ejJBjP8gvctqWhIR"
-    delete CONFIG.DND5E.armorIds.studded
-
-    //Shields
-    delete CONFIG.DND5E.shieldIds.shield
-    CONFIG.DND5E.shieldIds.large = "elkan5e.elkan5e-mundane-items.AENiTUeluTRiFzRz"
-    CONFIG.DND5E.shieldIds.small = "elkan5e.elkan5e-mundane-items.OE836KUoJiAsG0IA"
-}
