+++ lang/en.json
@@ -1,0 +1,14 @@
+{
+    "DND5E": {
+      "Feature.Race": "Ancestry Feature",
+      "ToolArtisans": "Artist's Tools",
+      "DamageThunder": "Sonic",
+      "ArmorClassDraconic": "Draconic Toughness"
+    },
+    "TYPES": {
+      "Item": {
+        "race": "Ancestry",
+        "racePl": "Ancestries"
+      }
+    }
+}
