+++ module.json
@@ -2,7 +2,7 @@
   "id": "elkan5e",
   "title": "Elkan 5e",
   "description": "Elkan Content",
-  "version": "1.2.6",
+  "version": "1.2.8",
   "compatibility": {
     "minimum": "11",
     "verified": "11.315"
@@ -225,8 +225,14 @@
       }
     ]
   },
-  "esmodules": [
-    "Javascript/elkan5e.js"
+  "esmodules": ["./Javascript/elkan5e.mjs" ],
+  "languages": [
+    {
+      "lang": "en",
+      "name": "English",
+      "path": "./lang/en.json",
+      "flags": {}
+    }
   ],
   "url": "https://github.com/etoy22/elkan5e",
   "manifest": "https://github.com/etoy22/elkan5e/releases/latest/download/module.json",
